# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  domain.BestSellerBook:
    properties:
      book_id:
        type: string
      title:
        type: string
      total_sold:
        type: number
    type: object
  domain.BookAuthor:
    properties:
      author_id:
        type: string
      author_name:
        type: string
      slug:
        type: string
    type: object
  domain.BookCategory:
    properties:
      category_id:
        type: string
    type: object
  domain.BookDetail:
    properties:
      authors:
        items:
          $ref: '#/definitions/domain.BookAuthor'
        type: array
      category:
        $ref: '#/definitions/domain.BookCategory'
      created_at:
        type: string
      detail_description:
        type: string
      id:
        type: string
      images:
        items:
          $ref: '#/definitions/domain.BookImage'
        type: array
      imported_at:
        type: string
      name:
        type: string
      price:
        type: number
      pricing:
        $ref: '#/definitions/domain.BookPricing'
      product_status:
        type: string
      series:
        $ref: '#/definitions/domain.BookSeries'
      short_description:
        type: string
      stock_quantity:
        type: integer
      tags:
        items:
          $ref: '#/definitions/domain.BookTag'
        type: array
    type: object
  domain.BookImage:
    properties:
      alt:
        type: string
      is_primary:
        type: boolean
      url:
        type: string
    type: object
  domain.BookListResponse:
    properties:
      books:
        items:
          $ref: '#/definitions/domain.BookDetail'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  domain.BookPricing:
    properties:
      price:
        type: number
    type: object
  domain.BookSeries:
    properties:
      sequence_no:
        type: integer
      series_id:
        type: string
      series_name:
        type: string
    type: object
  domain.BookTag:
    properties:
      tag_id:
        type: string
      tag_name:
        type: string
    type: object
...
  domain.DeactivateUserRequest:
    properties:
      is_active:
        type: boolean
    type: object
...
  domain.SalesSummary:
    properties:
      date_from:
        type: string
      date_to:
        type: string
      total_orders:
        type: integer
      total_revenue:
        type: number
    type: object
...
  domain.UserInfo:
    properties:
      alias_id:
        type: string
      email:
        type: string
      full_name:
        type: string
      phone:
        type: string
      role:
        $ref: '#/definitions/domain.UserRole'
    type: object
  domain.UserListResponse:
    properties:
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      users:
        items:
          $ref: '#/definitions/domain.UserInfo'
        type: array
    type: object
  domain.UserRole:
    enum:
    - user
    - admin
    type: string
    x-enum-varnames:
    - RoleUser
    - RoleAdmin
  server.errorResponse:
    properties:
      error:
        type: string
    type: object
  server.successResponse:
    properties:
      data: {}
    type: object
host: localhost:8080
info:
  contact:
    email: support@bookstore.local
    name: API Support
  description: Polyglot persistence bookstore backend — PostgreSQL · MongoDB · Neo4j
    · Redis.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: HCMUS Bookstore API
  version: "2.0"
paths:
  /admin/analytics/best-sellers:
    get:
      description: Return sales ranking data from Redis
      parameters:
      - description: Number of books
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.BestSellerBook'
            type: array
      security:
      - BearerAuth: []
      summary: 'Admin: Best sellers analytics'
      tags:
      - admin-analytics
  /admin/analytics/sales:
    get:
      description: Return total revenue and order count for a date range
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: from
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: to
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.SalesSummary'
      security:
      - BearerAuth: []
      summary: 'Admin: Sales summary'
      tags:
      - admin-analytics
...
  /admin/users:
    get:
      description: Return a paginated list of all registered users
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserListResponse'
      security:
      - BearerAuth: []
      summary: 'Admin: List users'
      tags:
      - admin-users
  /admin/users/{id}:
    get:
      description: Return profile data for any user
      parameters:
      - description: User Alias ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.UserInfo'
      security:
      - BearerAuth: []
      summary: 'Admin: Get user'
      tags:
      - admin-users
  /admin/users/{id}/deactivate:
    patch:
      consumes:
      - application/json
      description: Toggle the is_active flag for a user account
      parameters:
      - description: User Alias ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Activation status
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.DeactivateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.successResponse'
      security:
      - BearerAuth: []
      summary: 'Admin: Deactivate user'
      tags:
      - admin-users
...
  /books:
    get:
      consumes:
      - application/json
      description: Search and filter books from MongoDB catalog with live stock from
        PostgreSQL
      parameters:
      - description: Full-text search query
        in: query
        name: search
        type: string
      - description: Filter by author name
        in: query
        name: author
        type: string
      - description: Filter by publisher
        in: query
        name: publisher
        type: string
      - description: Filter by publish year
        in: query
        name: year
        type: integer
      - description: Minimum price
        in: query
        name: min_price
        type: number
      - description: Maximum price
        in: query
        name: max_price
        type: number
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Items per page (default 20)
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.BookListResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.errorResponse'
      summary: Search books
      tags:
      - books
  /books/{id}:
    get:
      consumes:
      - application/json
      description: Return full book metadata from MongoDB and stock from PostgreSQL
      parameters:
      - description: Book MongoDB ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.BookDetail'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/server.errorResponse'
      summary: Get book detail
      tags:
      - books
...
  /books/{id}/view:
    post:
      consumes:
      - application/json
      description: Increment daily view counter in Redis and log event in MongoDB
      parameters:
      - description: Book MongoDB ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/server.successResponse'
      security:
      - BearerAuth: []
      summary: Record book view
      tags:
      - books
  /books/new:
    get:
      consumes:
      - application/json
      description: Return the most recently imported books
      parameters:
      - description: Number of books to return (default 20)
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.BookDetail'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/server.errorResponse'
      summary: Get newest books
      tags:
      - books
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer " followed by your JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"